spaces_demo/requirements.txt
gradio
torch
transformers
Pillow